requirements.txt
django==2.2.1
djangorestframework==3.9.3
coverage==4.5.3
requests==2.21.0
django-cors-headers==2.5.3
django-tagging==0.4.6
django-hitcount==1.3.0
django-rest-auth==0.9.5
django-allauth==0.39.1
Pillow==6.0.0
django-debug-toolbar==1.11

